Rename lib utility files
